pyproject.toml
[tool.poetry]
name = "nautilus_trader"
version = "1.187.0"
description = "A high-performance algorithmic trading platform and event-driven backtester"
authors = ["Nautech Systems <info@nautechsystems.io>"]
license = "LGPL-3.0-or-later"
readme = "README.md"
homepage = "https://nautilustrader.io"
repository = "https://github.com/nautechsystems/nautilus_trader"
classifiers = [
    "License :: OSI Approved :: GNU Lesser General Public License v3 or later (LGPLv3+)",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering",
    "Topic :: Office/Business :: Financial",
    "Topic :: Office/Business :: Financial :: Investment",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "Intended Audience :: Science/Research",
]
include = [
    # Rust source must be included in the source distributions
    { path = "nautilus_core/*", format = "sdist" },
    { path = "nautilus_core/*[!target]*/*", format = "sdist" },
    # Compiled extensions must be included in the wheel distributions
    { path = "nautilus_trader/**/*.so", format = "wheel" },
    { path = "nautilus_trader/**/*.pyd", format = "wheel" },
    # Include the py.typed file for type checking support
    { path = "nautilus_trader/py.typed", format = "sdist" },
    { path = "nautilus_trader/py.typed", format = "wheel" },
    # Include Python interface files for type checking support
    { path = "nautilus_trader/**/*.pyi", format = "sdist" },
    { path = "nautilus_trader/**/*.pyi", format = "wheel" },
]

[build-system]
requires = [
    "setuptools",
    "poetry-core>=1.9.0",
    "numpy>=1.26.4",
    "Cython==3.0.8",
    "toml>=0.10.2",
]
build-backend = "poetry.core.masonry.api"

[tool.poetry.build]
script = "build.py"
generate-setup-file = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
cython = "==3.0.8"  # Build dependency (pinned for stability)
numpy = "^1.26.4"  # Build dependency
toml = "^0.10.2"  # Build dependency
click = "^8.1.7"
fsspec = "==2023.6.0"  # Pinned for stability
msgspec = "^0.18.6"
pandas = "^2.2.0"
pyarrow = ">=15.0.0"
pytz = ">=2023.4.0"
tqdm = "^4.66.1"
uvloop = {version = "^0.19.0", markers = "sys_platform != 'win32'"}

async-timeout = {version = "^4.0.3", optional = true}
betfair_parser = {version = "==0.10.0", optional = true}  # Pinned for stability
defusedxml = {version = "^0.7.1", optional = true}
docker = {version = "^7.0.0", optional = true}
nautilus_ibapi = {version = "==10.19.2", optional = true}  # Pinned for stability
pyjwt = "^2.9.0"

[tool.poetry.extras]
betfair = ["betfair_parser"]
docker = ["docker"]
ib = ["nautilus_ibapi", "async-timeout", "defusedxml"]

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "^24.1.1"
docformatter = "^1.7.5"
mypy = "^1.8.0"
pandas-stubs = "^2.1.4"
pre-commit = "^3.6.0"
ruff = "^0.2.1"
types-pytz = "^2023.3"
types-requests = "^2.31"
types-toml = "^0.10.2"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
coverage = "^7.4.1"
pytest = "^7.4.4"
pytest-aiohttp = "^1.0.5"
pytest-asyncio = "==0.21.1"  # Pinned due Cython: cannot set '__pytest_asyncio_scoped_event_loop' attribute of immutable type
pytest-benchmark = "^4.0.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
pytest-xdist = { version = "^3.5.0", extras = ["psutil"] }

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
numpydoc = "^1.6.0"
linkify-it-py = "^2.0.0"
myst-parser = "^0.18.1"
sphinx_comments = "^0.0.3"
sphinx_copybutton = "^0.5.0"
sphinx-external-toc = "^0.3.1"
sphinx-material = "^0.0.35"
sphinx_togglebutton = "^0.3.0"

[tool.isort]  # Used by legacy isort for Cython modules
py_version = "310"
skip_glob = ["**/core/rust/*"]
combine_as_imports = true
line_length = 100
ensure_newline_before_comments = true
force_single_line = true
include_trailing_comma = true
multi_line_output = 3
lines_after_imports = 2
use_parentheses = true
filter_files = true

[tool.black]
target_version = ["py310", "py311", "py312"]
line_length = 100

[tool.docformatter]
black = true
make-summary-multi-line = true
pre-summary-new-line = true
blank = true
recursive = true
in-place = true

[tool.ruff]
target-version = "py310"
line-length = 150  # Reduce to 100

exclude = [
    ".benchmarks",
    ".eggs",
    ".git",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    "build",
    "dist",
    "venv",
]

[tool.ruff.lint]
select = [
    "C4",
    "E",
    "F",
    "W",
    "C90",
    "D",
    # "DTZ",
    "UP",
    "S",
    "T10",
    "ICN",
    "PIE",
    # "PT",
    "PYI",
    "Q",
    "I",
    "RSE",
    "TID",
    # "SIM",
    # "ARG",
    # "ERA",
    "PD",
    # "PGH",
    # "PLW",
    "NPY",
    "RUF",
]

ignore = [
    "D100",  # Missing docstring in public module  **fix**
    "D101",
    "D102",  # Missing docstring in public method  **fix**
    "D103",  # Missing docstring in public function  **fix**
    "D104",  # Missing docstring in public package  **fix**
    "D107",
    "D105",
    "D200",  # One-line docstring should fit on one line with quotes (optional style)
    "D203",  # 1 blank line required before class docstring (optional style)
    "D205",  # 1 blank line required between summary line and description (optional style)
    "D212",  # Multi-line docstring summary should start at the first line (optional style)
    "D400",  # First line should end with a period (not always a first line)
    "D413",  # Missing blank line after last section ('Parameters')
    "D415",  # First line should end with a period, question mark, or exclamation point (not always a first line)
    "D416",  # Section name should end with a colon ('Warnings:', not 'Warnings') (incorrect?)
    "E741",  # Ambiguous variable name (single char)
    "PD901", # `df` is a bad variable name. Be kinder to your future self
    "RUF012",  # Mutable class attributes should be annotated with `typing.ClassVar`
    "S101",  # Use of assert detected (OK in test suite)
    "S105",  # Use of hardcoded password (spurious)
    "S106",  # Use of hardcoded password (spurious)
    "S113",  # Probable use of requests call without timeout **fix**
]

# Allow autofix for all enabled rules (when `--fix`) is provided
fixable = [
    "A",
    "B",
    "C",
    "C4",
    "D",
    "DTZ",
    "E",
    "F",
    "UP",
    "S",
    "W",
    "I",
    "PIE",
    "PT",
    "PYI",
    "RSE",
    "TID",
    "ARG",
    "PD",
    "SIM",
    # "PGH",
    "NPY",
    "RUF",
]

unfixable = []
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.isort]
force-single-line = true
single-line-exclusions = ["typing"]
lines-after-imports = 2

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.per-file-ignores]
"test_perf_logger.py" = ["S311"]

[tool.mypy]
python_version = "3.10"
disallow_incomplete_defs = true
explicit_package_bases = true
ignore_missing_imports = true
namespace_packages = true
warn_no_return = true
warn_unused_configs = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --new-first --failed-first --doctest-modules --doctest-glob=\"*.pyx\""
asyncio_mode = "strict"
filterwarnings = [
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
plugins = ["Cython.Coverage"]
source = ["nautilus_trader"]
omit = [
    "nautilus_trader/adapters/*",
    "nautilus_trader/examples/*",
    "nautilus_trader/test_kit/*",
]

[tool.coverage.report]
fail_under = 0
show_missing = true
